Move generate_pipeline_files to pipeline_writers.py
