+++ tests/unit/plugins/expected/airflow_pipeline_a0_b0/airflow_pipeline_a0_b0_Dockerfile
