+++ tests/unit/plugins/expected/airflow_pipeline_a0_b0_dependencies/airflow_pipeline_a0_b0_dependencies_Dockerfile
