+++ tests/unit/plugins/expected/script_pipeline_a0_b0/script_pipeline_a0_b0_module.py
