+++ tests/unit/plugins/expected/script_pipeline_a0_b0_dependencies/script_pipeline_a0_b0_dependencies_Dockerfile
