+++ tests/unit/plugins/expected/script_pipeline_housing_w_dependencies/script_pipeline_housing_w_dependencies_module.py
